added support for multiple addresses for a patient

+++ src/main/java/com/doctoranywhere/service/IPatientService.java
@@ -8,11 +8,12 @@ import com.doctoranywhere.model.Patient;
 public interface IPatientService {
 	List<Patient> findAllPatients();
 
-	Optional<Patient> findById(Long patientId);
+	Optional<Patient> findById(int patientId);
 
 	Patient save(Patient patient);
 
 	Patient update(Patient patient);
 
 	void delete(Patient patient);
+
 }
